feat(crypto): re-add `sha256` (hex)
